+++ src/views/Todos.vue
@@ -1,22 +1,25 @@
 <template>
   <div class="todos">
     <Label :initialValue="-1" :labels="labels" @update-label="onUpdateLabel" />
-    <Table :todos="todos" :labelValue="labelValue" />
+    <Table :todos="todos" :labelValue="labelValue" ref="table" />
+    <Form @do-add-form="onAddTodo" />
   </div>
 </template>
 
 <script lang="ts">
-import { defineComponent, ref } from "vue";
+import { defineComponent, ref, render } from "vue";
 import Table from "@/components/Table.vue";
 import { Todos, todoStorage } from "@/api/storage";
 import Label from "@/components/Label.vue";
 import { sampleLabels } from "@/api/label";
+import Form from "@/components/Form.vue";
 
 export default defineComponent({
   name: "Todos",
   components: {
     Table,
     Label,
+    Form,
   },
   setup: (props) => {
     // const labels = {};
@@ -28,17 +31,25 @@ export default defineComponent({
       labelValue.value = selected;
     };
 
+    const table = ref<InstanceType<typeof Table>>();
+
+    const onAddTodo = (comment: string) => {
+      table.value?.reactiveTodos.push({
+        id: todoStorage.uid++,
+        comment: comment,
+        status: 0,
+      });
+    };
+
     return {
+      table,
       todos,
       labels,
       labelValue,
       onUpdateLabel,
+      onAddTodo,
     };
   },
-  // props: {
-  //   todos: Array,
-  //   labels: Object,
-  // },
 });
 </script>
